+++ WaveBox/ApiHandler/Handlers/TestApiHandler.cs
@@ -13,7 +13,7 @@ namespace WaveBox.ApiHandler.Handlers
 		private HttpProcessor Processor { get; set; }
 		private UriWrapper Uri { get; set; }
 
-		public TestApiHandler(UriWrapper uri, HttpProcessor processor, int userId)
+		public TestApiHandler(UriWrapper uri, HttpProcessor processor, long userId)
 		{
 			Processor = processor;
 			Uri = uri;
